requirements.txt
Flask==2.0.2
Flask-CORS==3.1.1
google-api-python-client==2.0.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
gunicorn==20.1.0
gunicorn

